+++ src/Modules/Workflows/Domain/Engine/VariableResolvers/WorkflowResolver.php
@@ -46,6 +46,12 @@ class WorkflowResolver implements VariableResolverInterface
 
             case 'meta':
                 return new WorkflowMetaResolver($this->workflow['ID']);
+
+            case 'execution_id':
+                return (string)$this->workflow['execution_id'];
+
+            case 'execution_trace':
+                return (array)$this->workflow['execution_trace'];
         }
 
         return '';
@@ -53,14 +59,22 @@ class WorkflowResolver implements VariableResolverInterface
 
     public function getValueAsString(string $property = ''): string
     {
-        return (string)$this->getValue($property);
+        $value = $this->getValue($property);
+
+        if (is_array($value)) {
+            return implode(', ', $value);
+        }
+
+        return (string)$value;
     }
 
     public function compact(): array
     {
         return [
             'type' => $this->getType(),
-            'value' => $this->getValue('id')
+            'value' => $this->getValue('id'),
+            'execution_id' => $this->getValueAsString('execution_id'),
+            'execution_trace' => $this->getValueAsString('execution_trace'),
         ];
     }
 
@@ -83,6 +97,8 @@ class WorkflowResolver implements VariableResolverInterface
                 'description',
                 'modified_at',
                 'meta',
+                'execution_id',
+                'execution_trace',
             ]
         );
     }