+++ src/Modules/Workflows/Domain/Steps/Actions/Runners/AppendDebugLogRunner.php
@@ -3,9 +3,9 @@
 namespace PublishPress\Future\Modules\Workflows\Domain\Steps\Actions\Runners;
 
 use PublishPress\Future\Modules\Workflows\Interfaces\StepRunnerInterface;
-use PublishPress\Future\Modules\Workflows\Interfaces\RuntimeVariablesHandlerInterface;
 use PublishPress\Future\Framework\Logger\LoggerInterface;
 use PublishPress\Future\Modules\Workflows\Domain\Steps\Actions\Definitions\AppendDebugLog;
+use PublishPress\Future\Modules\Workflows\Interfaces\ExecutionContextInterface;
 use PublishPress\Future\Modules\Workflows\Interfaces\StepProcessorInterface;
 
 class AppendDebugLogRunner implements StepRunnerInterface
@@ -16,9 +16,9 @@ class AppendDebugLogRunner implements StepRunnerInterface
     private $stepProcessor;
 
     /**
-     * @var RuntimeVariablesHandlerInterface
+     * @var ExecutionContextInterface
      */
-    private $variablesHandler;
+    private $executionContext;
 
     /**
      * @var LoggerInterface
@@ -27,11 +27,11 @@ class AppendDebugLogRunner implements StepRunnerInterface
 
     public function __construct(
         StepProcessorInterface $stepProcessor,
-        RuntimeVariablesHandlerInterface $variablesHandler,
+        ExecutionContextInterface $executionContext,
         LoggerInterface $logger
     ) {
         $this->stepProcessor = $stepProcessor;
-        $this->variablesHandler = $variablesHandler;
+        $this->executionContext = $executionContext;
         $this->logger = $logger;
     }
 
@@ -61,7 +61,7 @@ class AppendDebugLogRunner implements StepRunnerInterface
                     $message = $message['expression'] ?? '';
                 }
 
-                $message = $this->variablesHandler->resolveExpressionsInText($message);
+                $message = $this->executionContext->resolveExpressionsInText($message);
                 $message = 'Slug: ' . $nodeSlug . ' | ' . $message;
                 $message = $this->stepProcessor->prepareLogMessage($message, $nodeSlug);