+++ src/Modules/Workflows/Domain/Steps/Actions/Runners/ScheduleDelayRunner.php
@@ -5,7 +5,7 @@ namespace PublishPress\Future\Modules\Workflows\Domain\Steps\Actions\Runners;
 use PublishPress\Future\Modules\Workflows\Domain\Steps\Actions\Definitions\ScheduleDelay;
 use PublishPress\Future\Modules\Workflows\Interfaces\AsyncStepProcessorInterface;
 use PublishPress\Future\Modules\Workflows\Interfaces\AsyncStepRunnerInterface;
-use PublishPress\Future\Modules\Workflows\Interfaces\RuntimeVariablesHandlerInterface;
+use PublishPress\Future\Modules\Workflows\Interfaces\ExecutionContextInterface;
 
 class ScheduleDelayRunner implements AsyncStepRunnerInterface
 {
@@ -15,16 +15,16 @@ class ScheduleDelayRunner implements AsyncStepRunnerInterface
     private $stepProcessor;
 
     /**
-     * @var RuntimeVariablesHandlerInterface
+     * @var ExecutionContextInterface
      */
-    private $variablesHandler;
+    private $executionContext;
 
     public function __construct(
         AsyncStepProcessorInterface $stepProcessor,
-        RuntimeVariablesHandlerInterface $variablesHandler
+        ExecutionContextInterface $executionContext
     ) {
         $this->stepProcessor = $stepProcessor;
-        $this->variablesHandler = $variablesHandler;
+        $this->executionContext = $executionContext;
     }
 
     public static function getNodeTypeName(): string
@@ -39,7 +39,7 @@ class ScheduleDelayRunner implements AsyncStepRunnerInterface
             function ($step) {
                 $nodeSlug = $this->stepProcessor->getSlugFromStep($step);
 
-                $this->variablesHandler->setVariable($nodeSlug, [
+                $this->executionContext->setVariable($nodeSlug, [
                     'schedule_date' => 0,
                     'is_recurring' => false,
                     'recurring_type' => '',