+++ src/Modules/Workflows/Domain/Steps/Triggers/Runners/OnInitRunner.php
@@ -19,6 +19,7 @@ class OnInitRunner implements TriggerRunnerInterface
      */
     private $logger;
 
+
     public function __construct(
         StepProcessorInterface $stepProcessor,
         LoggerInterface $logger