+++ src/Modules/Workflows/Module.php
@@ -229,7 +229,12 @@ final class Module implements InitializableInterface
     {
         // The execution_id variable is deprecated.
         if ($variableName === 'global.execution_id') {
-            return 'global.run_id';
+            return 'global.workflow.execution_id';
+        }
+
+        // The run_id variable is deprecated.
+        if ($variableName === 'global.run_id') {
+            return 'global.workflow.execution_id';
         }
 
         return $variableName;